rag_chatbot/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
tiktoken
